biblioteca/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==43.0.3
Django==5.1.3
django-ninja==1.3.0
django-ninja-extra==0.21.6
django-ninja-jwt==5.3.5
injector==0.22.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
sqlparse==0.5.2
typing_extensions==4.12.2
psycopg2-binary>=2.9.6, <2.10
